% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPro.R
\name{assignPro}
\alias{assignPro}
\title{Assignement of conditional probability values}
\usage{
assignPro(protree, evento, valPro)
}
\arguments{
\item{protree}{object of type proba_tree in which conditional probability values will be written.}

\item{evento}{a string representing an event of interest and further conditioning events.
For example 'Pa1:malato;Pa2:normal'}

\item{valPro}{conditional probability value.}
}
\value{
object of type proba_tree with assigned probability values
}
\description{
This is an important function to work with probability-event trees.
}
