% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condiziona.R
\name{condiziona}
\alias{condiziona}
\title{Conditionl probability of events of interest}
\usage{
condiziona(tavolaCongiunta, evento)
}
\arguments{
\item{tavolaCongiunta}{an object of type proba_tree where probability values have been already assigned.}

\item{evento}{a string describing the conditioned event of interest.}
}
\value{
conditional probability value.
}
\description{
Conditionl probability of events of interest
}
