% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congiungi.R
\name{congiungi}
\alias{congiungi}
\title{Joint probability of paths}
\usage{
congiungi(objprob)
}
\arguments{
\item{objprob}{an object of type proba_tree where probability values have been already assigned}
}
\value{
the joint distribution as defined by original partitions.
}
\description{
Joint probability of paths
}
