% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estraSottotavola.R
\name{estraSottotavola}
\alias{estraSottotavola}
\title{A helper to extract sub-tables}
\usage{
estraSottotavola(tavolaPro, evento)
}
\arguments{
\item{tavolaPro}{a full-sized table}

\item{evento}{event of interest}
}
\value{
sub-table
}
\description{
This is a function that extracts the subtables of interest.
}
