% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eve_tree.R
\name{eve_tree}
\alias{eve_tree}
\title{Object builder for type  eve_tree}
\usage{
eve_tree(myPart)

eve_tree(myPart)
}
\arguments{
\item{myPart}{a list of partitions for the sample space.}
}
\value{
the created object

the created object
}
\description{
This function crerates an object whose attribute Tipologia is  eve_tree.

This function crerates an object whose attribute Tipologia is  eve_tree.
}
