% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evtree_2ig.R
\name{evtree_2ig}
\alias{evtree_2ig}
\title{Convert an eve_tree object into an igraph object}
\usage{
evtree_2ig(primoPTree)
}
\arguments{
\item{primoPTree}{object of type eve_tree.}
}
\value{
an igraph  object
}
\description{
The igraph package is  needed for the trasformation taking place.
}
