% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchProb.R
\name{fetchProb}
\alias{fetchProb}
\title{A helper function to load probability values within an eve_tree object}
\usage{
fetchProb(oggettoMat, valoriProb)
}
\arguments{
\item{oggettoMat}{an object that contains probability values and tree paths.}

\item{valoriProb}{object that contains values to be inserted.}
}
\value{
an updated proba_tree
}
\description{
A helper function to load probability values within an eve_tree object
}
\examples{
\dontrun{
  # from partitions to probability values
  partiVac <- list(par1 =c("M","notM"),
  par2 = c("T1","notT1"))
  spazioCampio <-eve_tree(partiVac)
  primoPTree<- proba_tree(spazioCampio)
  primoPTree
  tmpGen <- genProbLs(primoPTree)
  tmpGen$probVal[ c(1,2)] <- c(0.0002, 1-0.0002)
  tmpGen$probVal[c(3,4)] <- c(0.99, 1-0.95)
  tmpGen$probVal[c(5,6)] <- c(0.01, 0.95)
  tmpGen[,-1]
  primoPTreefilled <- fetchProb(primoPTree,tmpGen)
  primoPTreefilled
  }


}
