% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genProbLs.R
\name{genProbLs}
\alias{genProbLs}
\title{Generate a list of all required assignements}
\usage{
genProbLs(objProT)

genProbLs(objProT)
}
\arguments{
\item{objProT}{Object of type 'proba_tree'}
}
\value{
a dataframe of required assignements; columns are path, event,
partition number, probability value.

a dataframe of required assignements; columns are path, event,
partition number, probability value.
}
\description{
Generate a list of all required assignements

Generate a list of all required assignements
}
