% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSimplex.R
\name{isSimplex}
\alias{isSimplex}
\title{Check is simplex}
\usage{
isSimplex(vettoval)

isSimplex(vettoval)
}
\arguments{
\item{vettoval}{a vector to be tested}
}
\value{
TRUE if the condition is sactisfied otherwise FALSE.

TRUE if the condition is sactisfied otherwise FALSE.
}
\description{
This function check if a collection of values adding to 1 is provided,
and in this case return TRUE.

This function check if a collection of values adding to 1 is provided,
and in this case return TRUE.
}
