% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalizza.R
\name{marginalizza}
\alias{marginalizza}
\title{Marginalization}
\usage{
marginalizza(tavolaPro, evento)

marginalizza(tavolaPro, evento)
}
\arguments{
\item{tavolaPro}{an object of conditional table where probability values are already assigned.}

\item{evento}{a string describing the event of interest.}
}
\value{
a probability value

a probability value
}
\description{
Marginalization

Marginalization
}
