% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proba_tree.R
\name{proba_tree}
\alias{proba_tree}
\title{Probabilistic tree}
\usage{
proba_tree(myETree)
}
\arguments{
\item{myETree}{an object of type "eve_tree" as create by
calling eve_tree()}
}
\value{
an object of type "proba_tree"
}
\description{
This  function  creates a matrix of conditional probabilities
given a dataframe of events.
}
